Make readme and some changes in tokens env and some_parsing.py
